+++ src/components/base/b-dynamic-page/b-dynamic-page.ts
@@ -48,9 +48,12 @@ import type {
 
 	iDynamicPageEl,
 	KeepAliveStrategy,
-	UnsafeBDynamicPage
+	UnsafeBDynamicPage,
+	OnBeforeRemovePage
 
 } from 'components/base/b-dynamic-page/interface';
+
+import ScrollManager from 'components/base/b-dynamic-page/modules/scroll-manager';
 
 export * from 'components/super/i-data/i-data';
 export * from 'components/base/b-dynamic-page/interface';
@@ -339,7 +342,8 @@ export default class bDynamicPage extends iDynamicPage {
 
 		const {
 			unsafe,
-			route
+			route,
+			r
 		} = this;
 
 		return new SyncPromise((resolve) => {
@@ -361,6 +365,8 @@ export default class bDynamicPage extends iDynamicPage {
 					currentPageComponent = currentPageEl?.component?.unsafe;
 
 				if (currentPageEl != null) {
+					r.emit('beforeRemovePage', <OnBeforeRemovePage>{saveScroll: ScrollManager.saveScrollIntoAttribute});
+
 					if (currentPageComponent != null) {
 						const
 							currentPageStrategy = unsafe.getKeepAliveStrategy(currentPage, currentRoute);
@@ -398,6 +404,10 @@ export default class bDynamicPage extends iDynamicPage {
 
 					if (pageComponentFromCache != null) {
 						pageComponentFromCache.activate();
+
+						unsafe.async.requestAnimationFrame(() => {
+							ScrollManager.restorePageElementsScroll(pageElFromCache);
+						}, {label: $$.restorePageElementsScroll});
 
 						unsafe.$el?.append(pageElFromCache);
 						pageComponentFromCache.emit('mounted', pageElFromCache);