+++ src/components/base/b-dynamic-page/interface.ts
@@ -71,3 +71,11 @@ export interface UnsafeBDynamicPage<CTX extends bDynamicPage = bDynamicPage> ext
 	// @ts-ignore (access)
 	wrapCache: CTX['wrapCache'];
 }
+
+export interface OnBeforeRemovePage {
+	/**
+	 * Saves the scroll position of an element
+	 * @param el
+	 */
+	saveScroll(el: Element): void;
+}