+++ src/components/base/b-dynamic-page/CHANGELOG.md
@@ -9,6 +9,13 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v4.0.0-beta.?? (2024-??-??)
+
+#### :rocket: New Feature
+
+* Added `beforeRemovePage` event that is emitted before removing the current page element
+* Added API for saving the horizontal scroll of nested DOM nodes on the page. [Learn more](./README.md#api-for-saving-scroll-of-nested-dom-nodes)
+
 ## v4.0.0-beta.49 (2024-01-17)
 
 #### :bug: Bug Fix